Introduce mlocate (version 0.26), replacing our current locate tool
This is packaged following the Fedora & Arch Linux model, where
/usr/bin/locate is a setgid executable, so it can read the
not-world-readable database file. The alternative model (found in
openSUSE) is to run as nobody and make the database world readable.

systemd .timer and .service files are provided, with the .service file
conditional on the state directory existing in the first place. That
way, the timer can be enabled by default but will do nothing until the
admin runs updatedb at least once.